+++ lib/postgres.dart
@@ -2,6 +2,7 @@ library postgres;
 
 export 'src/connection.dart';
 export 'src/execution_context.dart';
+export 'src/placeholder_identifier_enum.dart';
 export 'src/replication.dart' show ReplicationMode;
 export 'src/substituter.dart';
 export 'src/types.dart';